Setting up my DB
This is a push to my branch where I have the DBmodel updated to reflect the three tables, as well as, updated the web.config w/ my new connection string. I tested this once and added a new alarm to my DB which seemed to work out well, will be testing w/ tyler at 2:00pm today to ensure everything on my end is set up correctly.